BUG #17096: Fix missing fields in CGNS
vtkCGNSReader failed to correctly read the names of the solution nodes
when `FlowSolutionPointers` were present in the CGNS file. The reader is
expected to read the names for the Solution nodes from
`FlowSolutionPointers` data array. These names are often padded with
white spaces and hence lookup for the node using the solution node name
failed. Since CGNS docs specify that names cannot have whitespaces, we
now trim the solution node name read from `FlowSolutionPointers` to
remove any trailing whitespaces. That addresses part of #17096.