fix probe filter when using computed tolerance
There was an issue where when using a computed tolerance
it was using MAX_DOUBLE to lookup bins in the locator.
But the code that converts an XYZ to a bin overflows
on MAX_DOUBLE causing incorrect results. Specifically

void GetBucketIndices(const double *x, int ijk[3]) const

which happens when you pass to it a coordinate vector x with values
VTK_DOUBLE_MAX as argument
in that case the static_cast from double to vtkIdType
overflows causes incorrrect results.