Applied modifications about data transfer CPU/GPU
Following discussion : the pipeline has been corrected allowing it to
not process it as GPU -> CPU then CPU-> GPU. All is done with GPU
resulting in a much faster and more comfortable display in video mode.
To do so the architecture of the class has been slightly adapted to
needs.

Anyway, due to a probable issue (openvr issue #495), display is not
working with UndistortedFrameType but perfectly with others Frame Type.

For information and to investigate : since commit d7f9f0b9 "add a bunch
more debugging with VTK_REPORT_OPENGL_ERRORS", in debug mode, a lot of
OPENGL errors (those described in this commit changes) are triggered
just by opening the widget VR Menu for example.