correct incorrect axis labels under AdjustLabels option
Adjust labels automatically picks a set of "nice" numbers to use
as the axes labels and chooses new tick marks accordingly.
Unfortunately the placement of the new tick marks is incorrect
because they are scaled to the input data ranges instead of
the adjusted ranges. For instance if your top value was
35.5, you can end up with a tick labelled "40.0" at the 35.5
location.