correct incorrect axis labels under AdjustLabels option
"AdjustLabels" automatically picks a set of "nice" numbers to use
as the axes labels and chooses new tick marks accordingly.
Unfortunately the placement of the new tick marks is incorrect
because they are scaled to the input data ranges instead of
the adjusted ranges. For instance if your top value was
35.5, you can end up with a tick labelled "40.0" which is
great, except that it will be drawn at the 35.5 location,
which is not great.

(cherry picked from commit 0e3db7354ca8b2e91b2c273ee2f0822d99dde531)