vtkCGNSReader: Read only 32-char chunks from grid names
The CGNS data array `GridCoordinatesPointers` consists of 32-character
chunks, each containing the name of the coordinate grid to use for the
particular time step. ParaView was reading 33 characters from the list
and then setting the last character to NUL. This worked for all grids
except for the last one, where there are not enough characters for a
valid read-in of length 32. This commit fixes the indexing:

    offset + 1   for the start position
    offset + 32  for the end position

(CGNS uses Fortran indices.)