Reverse the logic for flipping TIFF images
The TIFF reader was flipping the image if the origin was not in the top-left,
but in my testing and reading the documentation vtkImageData's origin is in
the bottom-left. Comparing a number of TIFF images to those read by ImageJ,
Gwenview, and system previews to get a match in VTK tests or Tomviz requires
flipping the logic.