CGNS: avoid changing output based on ranks
vtkCGNSFileSeriesReader ended up changing the number of pieces in any
vtkMultiplePieceDataSet it generated based on how many ranks the reader
was reading it on. This was bad. This meant the composite index for any
block would change based on the run environment. Fixed that. Now the
vtkMultiplePieceDataSet will has exactly as many pieces as the splits in
the file.