Add scalar bar borders and tick marks + styling options
Adds a border around the color bar and swatches. This creates a visual
differentiation between the scalar bar and any grids of similar colors
the scalar bar may be overlapping.

Adds a background to annotation text. This allows for partially
transparent backgrounds to be applied to the text to increase contrast
and text readability.

Adds leaders to text labels to better connect the label with a specific
point on the scalar bar. A similar visual effect to how leaders are used
for annotations.

Minor corrections to sizing and positioning to improve spacing between
the scalar bar, swatches, and the border if not all swatches are enabled.