OffScreen Rendering of FastLabels
Added method to vtkGenericOpenGLResourceFreeCallback to determine
if a given vtkOpenGLRenderWindow is registered in it.

Modified vtkFastLabeledDataMapper::RenderPiece() to rebuild its
labels when the render window associated with it changes.

Before, if the render window changed after the labels were built
(e.g., an off-screen renderer is being used to export an image of
the current view), their texture handle would be released during
rendering and not be allocated again. Consequently, there would be
no labels in the produced image. To address this, render window
changes were included to the conditions of building the labels and
a reset of the associated render window is being done before building
them. This doesn't seem to cause any side effects to previous
fast labels functionality.