vtkExtractCells fixes
vtkExtractCells changed the cell ids filled by the user in its
RequestData. This was incorrect. It had the potential of dropping cell
ids provided is the input dataset had fewer elements for an earlier
iteration.

Also make vtkExtractCells use a `std::set` to store the cell ids. This
moves the cost of sorting / uniquifying from RequestData to setup.