Handle .vtu files with unsupported offset and cell types
Changes to vtkXMLUnstructuredGridReader restricted the offset and cell
DataArray types to int, long, and long long. However, the reader used
to support other DataArray types for these arrays and now produces an
error when attempting to read a file with this type offset/cell id.

This change checks whether the specified data array type is valid, and
if not, converts the data read from the file to an acceptable type
(long long to handle potentially large offsets and cell ids). Added
test for a non-supported offset and cell array type to verify the
conversion works.