Make all image GetDimensions() calls thread-safe
Previously, the Dimensions were computed from the Extent whenever
GetDimensions() was called.  This meant that the GetDimenions()
signature that returned the pointer to the Dimensions ivar was not
strictly thread-safe, since the ivar was written to on every call.

This change computes the Dimensions only when the Extent is set.
Now all GetDimensions() are thread-safe, eliminating a potential
source of thread-related bugs in VTK.