Nonmerging mode for edge filters
The vtkExtractEdges and vtkFeatureEdges filters now have a non-merging
mode which preserves the connectivity. Before it was only possible to
use a non-merging point locator which would create lots of duplicate
points and lose all connectivity information or a merging point locator
which would merge co-located points, thereby changing connectivity.