Add More CellMap Range Checks
When rotating, zooming, and interacting with a model, the
cellmap will rarely and randomly be empty even though the
cell map sizes will be properly calculated. By adding extra
guard conditions, for the frames which are not correctly
calculated we will not crash with the out of range accesses.

While not an ideal fix, as this code is called frequently enough,
and any issues will disappear with the next valid render, the
chances of a user getting a bad render and noticing is not likely.