Added new option to place and orient arrow source from the centre.
Arrow source tends to be placed a distance slightly off the centre of the entire arrow (depending on the length of its cylinder and tip) making scaling a bit off. Ideally having the arrow centralized in some cases makes it easier to achieve this.
Standalone Unit-test for ArrowSource class. (fixed comments)