New XML HTG V2 for reader and writer
This new format squashes together every arrays which used to be split by
tree. They can still be read by tree, given some new meta information on
the structure of the htg. One can still load an htg with a depth
delimiter.

Squashing the information saves a lot of space in the file. Prior to
that, one XML element was created per tree, which made outrageous large
files on htg with millions of trees.