Fix point data arrays in output data
Elements were being appended to point data arrays containing
additional vortex core identification criteria in an order that did
not match the order of the points. This was because points are added
to the output via a locator that merges coincident points, but
elements in the criteria array (for duplicate points) were just
appended to the end.

Fix this by creating a map from point IDs in the output to point IDs
in the criteria arrays. At the end of the algorithm, generate new
point data arrays that match the order of the points and add them to
the output instead of the original criteria arrays.