Update and use the VTK_QT_MAJOR_VERSION variable across Qt modules
This change adds a new CMake cache variable VTK_QT_MAJOR_VERSION for
determining the major version number of Qt used by VTK. This is an
internal variable, and users are not expected to set it. For now,
possible values for VTK_QT_MAJOR_VERSION are 5, 6.

By default, VTK looks for Qt6 and sets VTK_QT_MAJOR_VERSION to 6. If not
found, VTK looks for Qt5.