vtkDataAssemblyUtilities: reworking hierarchy
This commit changes how hierarchy is built. Now, when representating
a hierarchy, the dataset indices in the vtkDataAssembly generated are
the input dataset's composite indices.

This helps vtkDataAssemblyUtilities support a new API call that allows
us to convert selectors to composite ids either using a hierarchy or any
other data assembly associated with a partitioned dataset collection.