Update surface with edges
Update the approach to only draw pixels that are within the
triangle as opposed to outside of it. This fixes the overlay
lines that can show up with transparency and makes the shader
simpler and faster.

Downside is that the edges on non-manifold surfaces are half width
when previously they were closer to whole width.

See https://gitlab.kitware.com/vtk/vtk/-/issues/18064