clang-tidy: simplify bool return statements
This commit resolves `readability-simplify-boolean-expr` lints where a
bool is returned from a function and a conditional is used to decide the
final return value.

This does not resolve all instances, but just those which are "complete"
versus functions where other logic may occur in between the last
conditional and a default return statement.