Adding a smooth transition in min and max curv
H : mean curvature - K : gaussian curvature
When H^2 - K < 0 in vtkCurvatures, k_min and k_max (min and max
curvatures) are undefined. This undefinition comes from a small to large
error in the gaussian and mean curvature estimation.

If H^2 - K ~ 0, there is no reason to just set k_min and k_max to zero,
as its value is more like, for instance, k_min = H + i*something_small.

This commit generates a smooth transition so that H^2 - K ~ 0 outputs H
for k_min and k_max, and 0 when K >> 0. A gaussian kernel is used to
define this window, with a standard deviation set to 1e-3