Support reading separate boundaries in files
Previously, surfaces in the datasets were collected into a single
vtkPolyData object. This ignored the fact that there can be separate
surface objects within a CONVERGE file, each with a separate name.

This change supports reading multiple surfaces from a file and stores
the surfaces as separate vtkPolyData objects in a
vtkMultiBlockDataSet.

Updated test to reflect the new multiblock structure produced by the
reader.

Co-author: Paul Gruenbacher <paul.gruenbacher@convergecfd.com>