Make a list of headers that cannot be found
This allows us to avoid hitting the filesystem repeatedly when a
header file is not in any of the include directories.  It is also
useful for diagnostics, since it allows us to track such headers.

The "missing" headers are C and C++ standard library headers that
are kept in directories that the compiler knows about, but which
are unknown to the wrapper tools.