Cleanup some use of depth textures in OpenGL2
VTK classes to Blits between depth buffers which can be
problematic as blits between different formats is an error.

As VTK mappers and passes have to often work inside of an unknown
FBO add some methods to help create depth textures that are
compatible with the currently bound depth format.

There are a number of render pasees that have a DepthFormat ivar
that should also be updated to default to creating a compatible
format to be safe. This is especially true though of classes that
perform depth blits.