Add vtkFeatureEdges::PassLine + bug fix
When this new flag is on, input lines in a poly data are passed to the
output. This is the only way they can be passed.

This commit also fixes a bug: when the input has mixed cell types
(Strips, Verts, Polys, Lines), the mapping from the input cell ids to
the output cell ids was wrong. The multiplicity of input cell arrays was
not taken into account. This is now fixed

Bug fix:
Cell Data was wrongly mapped in the output when the input had mixed
types. This came from the fact that cell are sorted per main types in a
vtkPolyData. This commit fixes this bug and adds a test ensuring that
cells are rightly mapped.