Fewer fast floor impls for image interpolators
There were many many checks for specific CPU architectures, many
of which are of only historical interest.  The new code has only
three implementations: one for 64-bit CPUs, one for little-endian
32-bit CPUs, and one for big-endian 32-bit CPUs (the latter two
implementations are identical apart from the endianness).

The previous fallaback that used vtkMath::Floor() would give the
same integer from floor, but a slightly different fractional
remainder.  The difference was at most 2^(-16), and therefore not
generally significant, but the this fallback was superfluous
since the three implementations described above are sufficient.