Reduce compile time and memory for vtkDataArray
The DoComputeScalarRange() templated method was declared extern,
but that's apparently not enough for gcc to avoid processing
the template code as if the instantiations were occurring in the
current translation unit.  In order to reduce the workload on the
compiler, it's necessary to actually hide the templated method
definitions from the compiler.