Ghost cells generator: poly data bug fix for cell data
Cell data was wrongly mapped when generating ghost cells on a poly data.
This was not being caught because of an error in the test file. This
error is fixed and the bug as well.

The test also tests cell data when one input partition already having
ghosts.

Addresses paraview/paraview#20860