Add UnitTestCells test for vtkBiQuadraticPyramid
Apart from adding a test case for vtkBiQuadraticPyramid in
UnitTestCells, a lot of initializations are now using `auto`.
Additionally, some mid-edge points coordinates of the
vtkQuadraticPyramid were fixed, because they were not actually
the mid-edge points. Lastly, some `MakeCell` functions are now
using the same functions as in vtkCellValidator for consistency.