vtkBuffer: realign `DeleteFunction` when using `MallocFunction`
This can occur when the buffer is reallocating memory that was passed in
from the outside with a different delete method. Since `MallocFunction`
is now being used, `DeleteFunction` may need to be reset to match with
it.

Really, this points to the fact that tracking where the data backing the
buffer was actually allocated and keeping track of externally-provided
arrays using yet another `DeleteFunction` should be performed. This can
be work for the future.