Change incorrect fix to lint warning
The fix for a lint warning in a4d1bfc1 incorrectly invoked the parent
class method instead of the grandparent's method. This led to a slew
of error messages in vtkCompositeControlPointsItem whenever the number
of control points in the opacity function exceeded the number of
control points in the color transfer function. Fix by invoking the
grandparent's method as before the lint fix, but in a way that shouldn't
trigger a lint warning.