vtkDescriptiveStatistics: fixing MPI + overflow
`vtkPDescriptiveStatistics` was generating the statistics result on all
ranks, resulting on duplicated data on distributed inputs. Now on, only
rank 0 keeps statistics.

The only impacted output port is port 1, which delivers the output model.
For any rank but rank 0, it is now composed of a `vtkMultiBlockDataSet`
with same structure as in rank 0, but with empty leaves.

A test is being added to ensure that the filter has the correct
behavior. Numerical instability caused by overflowing ids caused
computation errors. This is now fixed.

Kurtosis formula was wrong for distributed input. The formula is now
fixed and its new origin is documented both in doxygen and in the code.

Dead test TestRealPDescriptiveStatisticsMPI.cxx is being removed

Addresses paraview/paraview#20854