cell locators: compatibility for polyhedra-polygon
The method `vtkAbstractCellLocator::FindCell(double[3])` needs to
allocate a `weight` array for the subrouting to work. The size of this
array was hardcoded to be 32. In the instance of polyhedra and polygons,
the weight array can need to be arbitrarely larger.

`vtkAbstractCellLocator::BuildLocator()` now has an implementation that
updates the array size if the input is an unstructured grid or a poly
data with polyhedrons or polygons.