vtkTransform: fix the return type of GetConcatenatedTransform
The docs indicate that it was intended to be a `vtkAbstractTransform`,
but it was declared as a `vtkLinearTransform` and casted into a
`vtkLinearTransform*`. However, AddressSanitizer detected that it is not
actually an instance of `vtkLinearTransform`, so using it as such a
pointer was undefined behavior.

While this is an API break, any code relying on the old code is not
actually working as intended since the object does not actually have
access to `vtkLinearTransform` methods.