OpenXR: add OpenXRModel and reparent OpenXRCamera
Added the OpenXRModel class and updated the places that were
using VRRay (as a placeholder) to instead use model as they should.

Implemented a very basic model that just draws a green pyramid.
Eventually we need a better solution. There is an extension
that is supported on Windows mixed reality that can provide
the controller model as a gltf2 glb but it doesn't seem to
be available on steamvr yet. Stub added to show how this could be
used.

Reparented OpenXRCamera to be a child of VRCamera and reworked
the transformations so that the TCDC matrices were available
for the model to use in rendering.