vtkIdList: adding memory management mechanism.
It was not possible to create a `vtkIdList` that would just read a
subset of another `vtkIdList`, because each `vtkIdList` owns its
internal buffer.

A mechanism to optionally disown the internal buffer is provided in
`vtkIdList::SetArray`. From now on, one can assign a pointer and a given length and
tell `vtkIdList` that someone else is managing the pointer.