Ghost cells generator: input global ids fix for ug
The global point ids were not propagated properly along the pipeline
when the input was an unstructured grid (it worked for poly data).
This is because the surface filter being used doesn't copy them in the
output.

This is fixed by untagging the global ids of the input in a temporary
data set and then manually fetching the array into the local block
structure.

Adresses paraview/paraview#21067