Ghost cells generator: preprocess with bounding box fix
When weeding out points using `vtkBoundingBox::ContainsPoint`, if the
point was slightly outside of the bounding box, then it was possible to
not keep in one block, but not in the other. Adding a tolerance parameter to this
function helps keep those points around.

The same problem was occuring with the usage for
`vtkBoundingBox::Intersects`.