vtkDataSetAttribute::CopyData ambiguity fix
The templated version of the method `vtkDataSetAttribute::CopyData`
aiming at avoiding ambiguity was not thorough.

`std::is_scalar<vtkIdList*>` returns true. To discriminate such
pointers, `!std::is_pointer<vtkIdList*>` needs to be added.
We need to keep `std::is_scalar` because
`std::is_scalar<vtkNew<vtkIdList>>` is a thing.
Moreover, in order to avoid infinite recursions, `vtkIdType` inputs
need weeded out in the template version of the method

The method is shadowed in subclasses `vtkCellData` and `vtkPointData`,
so `using Superclass::CopyData` is needed.