Add Custom Filtered Data Output Types
By allowing a user to select their own output type, the filter can be
configured to filter polydata into polydata instead of forcing the
output to be an unstructured grid.

This is important for v9.1+ as some point data can no longer be
reconstructed to polydata using a geometry filter.