Test Fast Labels
The fast labels test has been rewritten so it can be the basis of an
automated test to verify the rendering against regressions. It also
has a few new features to make it easier for a user to manually run
and test the interactivity of the mapper.

The fast label mapper also has a bugfix so it detects changes to strings
which don't have any new characters used, but change one character
for another.