vtkExecutive: break the reference cycle with `vtkAlgorithm`
This is done by only holding a raw pointer from `vtkExecutive`.
Eventually, this should be a weak pointer of some kind (see !8686), but
until there is a safe one, a raw pointer is sufficient.

In order to ensure that subclasses do not interfere with the invariants,
the `Algorithm` member is made private and access is through
`GetAlgorithm` instead.