Raise an error if external data store is needed
The default location SOURCE_DIR/.ExternalData for external data exists in
the VTK git repository but not in the release tarballs.  For builds from
tarballs, cmake used to silently turn off VTK_BUILD_TESTING because of this.
Now, cmake raises FATAL_ERROR and asks for an external data directory to be
specified.