vtkTemporalInterpolatedVelocity: Support LinearTransformation/SameTopology
vtkTemporalInterpolatedVelocity was using vtkCachingInterpolatedVelocityField.
vtkCachingInterpolatedVelocityField was an old class which can now be replaced by
vtkCompositeInterpolatedVelocityField. By doing that we can select a find cell strategy.

Also, vtkParticleTracerBase along with vtkTemporalInterpolatedVelocity they support datasets
that at each time-step they are a linear transformation of the first time-step, or they have the same
SameTopology.