Rewrite vtkProbeLineFilter algorithm
This new algorithm is now able to support mixture of 2D cells, 3D cells,
as well as holes along the line, which was not the case before.
It does not use vtkPProbeFilter for the sampling patterns SEGMENT_CENTERS
and CELL_BOUNDARIES but directly tries to find intersected cells with locators
and compute interpolated attributes using the parametric coordinates of the
intersection point instead, so computed positions and values are more robust.