fix(alpha-blending): fix `clearAlpha()` to support clearing the alpha buffer
``Qt 5.6`` switched from `straight alpha blending mode only`_ to
`straight or pre-multiplied`_. This change was accompanied by several
regressions for libraries and ``Qt`` widgets that previously used straight
alpha blending:

    - `QTBUG-74285`_ and fix `Qt 255297`_
    - `QTBUG-55245`_ and fix `Qt 168873`_
    - `vispy #1889`
    - `mrtrix #266`_
    - `paraview #17159`_

This method adopts the solution discovered in the previous issues of clearing
the alpha buffer and returning the correct value.

Unfortunately, in ``PySide6 <=6.3```, several ``glGet`` functions from
``QOpenGLFunctions`` are missing or broken:

    - [PYSIDE-2013](https://bugreports.qt.io/browse/PYSIDE-2013)
    - [PYSIDE-2017](https://bugreports.qt.io/browse/PYSIDE-2017)

Thus, we do not restore the previous OpenGL state before clearing the
buffer. This is generally not a problem since typically only ``VTK`` is
rendering in the viewport. However, if ``Qt`` widgets are overlayed in
the viewport, the user may experience issues. Detailed instructions for
tackling this issue are provided in the method docstring.

.. _`straight alpha blending mode only`:
    https://github.com/qt/qtbase/blob/4a1e5dbade4bab55f39bd368480dcca9a11e4b38/src/gui/painting/qplatformbackingstore.cpp#L264

.. _`straight or pre-multiplied`:
    https://gitlab.kitware.com/vtk/vtk/-/blob/master/Rendering/OpenGL2/vtkOpenGLRenderWindow.cxx#L1254

.. _`QTBUG-74285`:
    https://bugreports.qt.io/browse/QTBUG-74285

.. _`Qt 255297`:
    https://codereview.qt-project.org/c/qt/qtbase/+/255297

.. _`QTBUG-55245`:
    https://bugreports.qt.io/browse/QTBUG-55245

.. _`Qt 168873`:
    https://codereview.qt-project.org/c/qt/qtbase/+/168873

.. _`vispy #18891`:
    https://github.com/vispy/vispy/issues/1889

.. _`mrtrix #266`:
    https://github.com/MRtrix3/mrtrix3/pull/266

.. _`paraview #17159`:
    https://gitlab.kitware.com/paraview/paraview/-/issues/17159